Implement "Border + Fill" tiles.
This code seems very weird and janky but the thing is that this problem is simply quite annoying. You have 256 cases to consider and there doesn't seem to be a nice way to sift through those cases without doing it manually. Eventually you get down to 47(prime number) cases and you just have to manually assign them to the correct texture coordinates.